# license-application-service/src/main/resources/api.yaml
openapi: 3.1.0
info:
  title: Mallorca Holiday Rental Licensing API
  version: 1.0.0
  description: |
    API for managing applications, licenses, and payments for the Mallorca Holiday Rental Licensing system.
    Supports GDPR-compliant user management through Keycloak integration and ensures transparent, fair allocation of ETV licenses.
servers:
  - url: https://localhost:8080/
...
security:
  - KeycloakOAuth2:
      - openid
      - profile
      - email
paths:
  /login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate a user and obtain JWT access and refresh tokens.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResource"
        "401":
          description: Unauthorized - invalid credentials
  /register:
    post:
      tags: [Authentication]
      summary: User registration
      description: Register a new user account in the system.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResource"
        "400":
          description: Bad request - invalid input data
        "409":
          description: Conflict - user already exists
  /refresh-login:
    post:
      tags: [Authentication]
      summary: Refresh JWT tokens
      description: Refresh the JWT access token using a valid refresh token.
      operationId: refreshLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  description: The refresh token issued during login.
                  example: dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4...
      responses:
        "200":
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResource"
        "401":
          description: Unauthorized - invalid or expired refresh token
  /reset-password:
    post:
      tags: [Authentication]
      summary: Request password reset
      description: Initiate a password reset process for a user by email.
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset email sent
        "404":
          description: User with the given email not found
  /change-password:
    post:
      tags: [Authentication]
      summary: Change user password
      description: Change the password for an authenticated user.
      operationId: changeUserPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "201":
          description: Password changed successfully
        "400":
          description: Bad request - invalid input data
  /change-userdetails:
    post:
      tags: [ Authentication ]
      summary: Change user details
      description: Change the lastname or the email of a user
      operationId: changeUserDetails
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeUserdetailsRequest"
      responses:
        "201":
          description: Details changed successfully
        "400":
          description: Bad request - invalid input data
        "409":
          description: Conflict - email already in use
...
components:
  securitySchemes:
    KeycloakOAuth2:
      type: oauth2
      description: OAuth2 with Keycloak JWT Bearer tokens.
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8080/auth/realms/license-realm/protocol/openid-connect/auth
          tokenUrl: http://localhost:8080/auth/realms/license-realm/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect scope
            profile: Profile information access
            email: Email access
  schemas:
    LoginRequest:
      type: object
      description: Request payload for user login.
      properties:
        email:
          type: string
          format: email
          description: User's username.
          example: johndoe@email.com
        password:
          type: string
          format: password
          description: User's password.
          example: securepassword123
      required:
        - username
        - password
    LoginResource:
      type: object
      description: Response payload containing authentication tokens.
      properties:
        access_token:
          type: string
          description: JWT access token for authenticated requests.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..
          x-field-extra-annotation: '@JsonProperty("access_token")'
        refresh_token:
          type: string
          description: Token used to obtain a new access token.
          example: dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4...
          x-field-extra-annotation: '@JsonProperty("refresh_token")'
        expires_in:
          type: integer
          description: Time in seconds until the access token expires.
          example: 3600
          x-field-extra-annotation: '@JsonProperty("expires_in")'
        refresh_expires_in:
          type: integer
          description: Time in seconds until the refresh token expires.
          example: 86400
          x-field-extra-annotation: '@JsonProperty("refresh_expires_in")'
        token_type:
          type: string
          description: Type of the token issued.
          example: Bearer
          x-field-extra-annotation: '@JsonProperty("token_type")'
        is_admin:
          type: boolean
          description: Indicates if the user has administrative privileges.
          example: false
      required:
        - access_token
        - refresh_token
        - expires_in
        - token_type
    RegisterRequest:
      type: object
      description: Request payload for user registration.
      properties:
        firstname:
          type: string
          description: User's first name.
          example: John
        lastname:
          type: string
          description: User's last name.
          example: Doe
        email:
          type: string
          format: email
          description: User's email address.
          example: johndoe@email.com
        password:
          type: string
          format: password
          description: User's desired password.
          example: securepassword123
    RegisterResource:
      type: object
      description: Response payload after successful user registration.
      properties:
        user_id:
          type: string
          format: uuid
          description: Unique identifier of the newly created user.
          example: f28d1d3b-9bcb-4a74-a65a-2fede2b0a6c3
        message:
          type: string
          description: Confirmation message.
          example: User registered successfully. Please verify your email to activate the account.
    ChangeUserdetailsRequest:
      type: object
      description: Request payload for user details change.
      additionalProperties: false
      minProperties: 2
      properties:
        userId:
          type: string
          format: uuid
          description: User's unique identifier.
          example: f28d1d3b-9bcb-4a74-a65a-2fede2b0a6c3
        firstname:
          type: string
          description: User's new first name.
          example: John
        lastname:
          type: string
          description: User's new last name.
          example: Doe
        email:
          type: string
          format: email
          description: User's new email address.
          example: johndoe@email.com
      required:
        - userId
    ResetPasswordRequest:
      type: object
      description: Request payload for initiating a password reset.
      properties:
        email:
          type: string
          format: email
          description: Email address associated with the user account.
          example: john.doe@example.com
    ChangePasswordRequest:
      type: object
      description: Request payload for changing the user's password.
      properties:
        token:
          type: string
          format: bytes
          description: User's token.
          example: dmFoYW4uZHVjaGVyQGljbG91ZC5jb206MTc2Nzk1ODU1OTE4MQ==
        new_password:
          type: string
          format: password
          description: User's new desired password.
          example: newsecurepassword456
...